pyproject.toml
[project]
name = "news-summarizer-p"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-data-tables>=12.7.0",
    "azure-functions>=1.24.0",
    "azure-storage-blob>=12.27.1",
    "beautifulsoup4>=4.14.2",
    "bs4>=0.0.2",
    "openai>=2.8.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sendgrid>=6.12.5",
]
